Create Looping
Write a program for a small store that sells food and beverages.The program will ask the user if the user wants to order.
If yes, the program will provide categories for the user for the items he/she wishes to buy such as :
Food: (add as many food you want for the list)
 Snacks: (add as many snacks you want)
  Beverages: (add as many beverages you want)
As long as the user wish to order, the program will keep on providing options to buy.
When the user decides not to order anymore, the program will list the items the user ordered.
You decide for the prices of each item. Display the total of all the items the user ordered.